+++ server/data_access_layer/daos/chart_report.dao.js
@@ -1,7 +1,6 @@
 const databases = require("../databases");
 const ChartReportModel = databases['localdb'].chartReports;
 const ChartReportDataModel = databases['localdb'].chartReportsData;
-const PerformanceReportModel = databases['localdb'].performanceReport;
 
 // Chart Report Related functions
 exports.getChartReportsByUserId = async (userId, chartReportModel = ChartReportModel) => {
@@ -95,25 +94,6 @@ exports.deleteChartReportById = async(chartReportId, chartReportModel = ChartRep
                     status: err.status || 500,
                     message: err.message || "Could not delete data."
                 }
-                reject(response);
-            });
-    });
-}
-
-// Reports Types related functions
-exports.getPerformanceReportsWhenConnectedAsAdmin = async (userId, performanceReportModel = PerformanceReportModel) => {
-    return new Promise((resolve, reject) => {
-        performanceReportModel.findAll().then(async data => {
-            if (data) {
-                resolve(data)
-            }
-            resolve(false);
-        })
-            .catch(async err => {
-                const response = {
-                    status: err.status || 500,
-                    message: err.message || "Could not fetch data."
-                };
                 reject(response);
             });
     });
